S15.e.5: extract operator.ts cluster + topology render (1880 → 1586 LOC)
Fifth sub-slice of the S15.e operator.ts decomposition train.

Lifts renderTopology (with its nested makeBox/fitVis/visualLen/
statusIcon helpers) and renderCluster (with makeBar) out of the giant
startDashboard closure into a new operator/render/ module group:

cli/src/commands/operator/render/topology.ts (~199 LOC):
  - renderTopology(ctx) — closure deps: sandboxes, securityStates, topologyBox

cli/src/commands/operator/render/cluster.ts (~143 LOC):
  - renderCluster(ctx) — closure deps: clusterData, clusterNodeBox, clusterInfoBox
  - makeBar(pct) — pure helper

Two thin wrappers remain in operator.ts so call sites in the
orchestrator render() are unchanged.

Behavioral change: NONE.
- Bodies byte-identical to originals.
- Closure-captured state injected via RenderContext interfaces.
- sandboxes array reassigned in refresh(); wrapper re-reads on each
  call, preserving original semantics.

Removed now-unused imports: NodeInfo, sumPrometheusCounter.

Cumulative S15.e delta:
  pre-S15.e: 2894
  e.1:       2739 (-155)
  e.2:       2483 (-256)
  e.3:       1960 (-523)
  e.4:       1880  (-80)
  e.5:       1586 (-294)
  Cumulative: -1308. Heavy renders + dialogs (~1400 LOC) remaining for cap.

Verification:
- tsc --noEmit clean / lint 20 warnings (was 21) / build clean
- npm test: 454 pass / 2 skipped pre-existing

Audit: docs/security-audits/2026-04-29-phase2-hotspot-operator-cli-e5.md
